Fix getEighths method of Direction class

+++ src/main/java/com/snek/frameworklib/data_types/graphics/Direction.java
@@ -34,7 +34,7 @@ public enum Direction {
         return name;
     }
 
-    public double getEighths() {
+    public int getEighths() {
         return eighths;
     }
